first cut at parse_query
Code can now tell which components (SELECT, FROM, WHERE) are in user query. And can tell at which index in the query string the component starts. Next is to grab the value of the query component into query_term_values dict. (Then applying the query to the csv files of course.)